requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
pydantic==2.13.3
httpx==0.27.2
python-dotenv==1.0.1